Extract toLocaleDateObject() to plan entity
